:truck: rename auth-dir
