+++ aksel.nav.no/website/tsconfig.json
@@ -20,6 +20,7 @@
       "@/sb-util": ["./.storybook/utils.tsx"],
       "@/types": ["./components/types/index.ts"],
       "@/sanity-block": ["./components/website-modules/SanityBlockContent.tsx"],
+      "@/auth/*": ["./components/auth/*"],
       "@/hooks/*": ["./components/hooks/*"],
       "@/preview": ["./components/website-modules/preview/WithPreview.tsx"],
       "@/logging": ["./components/logging/index.ts"],
